// 3ME_UI/src/components/pickUpAndDischargeRequest/pickUpRequest/pickUpRequestPageSection.enum.ts
export enum PickUpRequestPageSection {
  PICK_UP_REQUEST_START_FORM = "pickup-start-form",
  PICK_UP_REQUEST_SUBMIT_FORM = "pickup-submit-form",
  PICK_UP_REQUEST_CONFIRM_FORM = "pickup-confirm-form",
  DISCHARGE_REQUEST_START_FORM = "discharge-start-form",
  DISCHARGE_REQUEST_SUBMIT_FORM = "discharge-submit-form",
}